// packages/client/src/components/game_engaine/Settings.ts
// Настройки для игры

enum Settings {
  // Ширина экрана
  ScreenWidth = 400,

  // Высота экрана
  ScreenHeight = 500,

  //Ширина трубы
  PipeWidth = 65,

  //Высота трубы
  PipeHeight = 150,

  // Зазор между труб
  PipeGap = 120,

  // Гравитация
  Gravity = 0.98,

  // Скорость птицы
  SpeedBird = 0,

  // Ширина птицы
  BirdWidth = 45,

  // Высота птицы
  BirdHeight = 35,
}

export default Settings